requirements.txt
ithaca-py
python-dotenv
rich
taipy